think_big_2023/requirements.txt
discord.py==2.2.2
Pillow==9.5.0
requests==2.28.2

# Ai
openai==0.27.4
python-dotenv==1.0.0
watchdog==3.0.0


# Tasks
redis==4.5.4
celery==5.2.7
langchain==0.0.138
tiktoken==0.3.3


# Youtube Summarizer
pytube==12.1.3
